Split up monitor-update-failed tests and other functional tests
