+++ src/ln/mod.rs
@@ -28,3 +28,5 @@ mod onion_utils;
 #[macro_use] mod functional_test_utils;
 #[cfg(test)]
 mod functional_tests;
+#[cfg(test)]
+mod chanmon_update_fail_tests;
